extensions/project-soul/soul/emotion.py
"""
Emotion detection and response modulation
"""

from typing import Dict, Any
from dataclasses import dataclass
from loguru import logger

from .config import EmotionConfig
from .moshi_client import AudioParams
from .liaison import TaskResult


@dataclass
class Emotion:
    """Detected emotion state"""
    name: str  # "frustrated", "excited", "confused", etc.
    confidence: float  # 0.0 - 1.0
    intensity: float  # 0.0 - 1.0


class EmotionEngine:
    """
    Detect user emotion from voice and modulate Soul's responses

    NOTE: This is a placeholder implementation. Full emotion detection requires:
    1. Audio feature extraction (librosa, pyannote.audio)
    2. Emotion classification model
    3. Voice prosody analysis (pitch, tempo, energy)
    """

    def __init__(self, config: EmotionConfig):
        self.config = config
        logger.info("Initializing emotion engine")
        logger.warning(
            "⚠️  Emotion detection is not yet implemented. "
            "This is a placeholder for future implementation."
        )

    async def detect_user_emotion(self, audio: bytes) -> Emotion:
        """
        Analyze user's voice for emotional state

        Args:
            audio: Audio bytes from user

        Returns:
            Detected emotion

        TODO: Implement actual emotion detection
        This should:
        1. Extract prosody features (pitch, tempo, energy)
        2. Classify emotion using trained model
        3. Consider conversation context
        4. Track emotional trajectory
        """
        # Placeholder: return neutral
        return Emotion(
            name="neutral",
            confidence=1.0,
            intensity=0.5,
        )

    async def modulate_response(
        self,
        result: TaskResult,
        user_emotion: Emotion,
    ) -> AudioParams:
        """
        Generate appropriate response parameters based on user emotion

        Args:
            result: Task result to communicate
            user_emotion: Detected user emotion

        Returns:
            Audio parameters for Moshi

        TODO: Implement sophisticated response modulation
        This should:
        1. Match or complement user emotion
        2. Consider task outcome (success/failure)
        3. Apply empathy when needed
        4. Celebrate successes appropriately
        """
        if not self.config.enable_response:
            return AudioParams()

        # Get emotion config
        emotion_config = self.config.emotions.get(
            user_emotion.name,
            {"speed": 1.0, "pitch": 1.0, "add_pause": False}
        )

        # Adjust based on task result
        if not result.success and user_emotion.name == "frustrated":
            # User is frustrated and task failed - be extra calm
            return AudioParams(
                tone="empathetic",
                speed=0.85,
                pitch=0.95,
                add_pause=True,
            )
        elif result.success and user_emotion.name == "excited":
            # User is excited and task succeeded - celebrate!
            return AudioParams(
                tone="enthusiastic",
                speed=1.15,
                pitch=1.1,
                add_laughter=True,
            )
        else:
            # Use configured emotion response
            return AudioParams(
                tone="supportive",
                speed=emotion_config.get("speed", 1.0),
                pitch=emotion_config.get("pitch", 1.0),
                add_pause=emotion_config.get("add_pause", False),
                add_laughter=emotion_config.get("add_laughter", False),
            )

    def _extract_prosody(self, audio: bytes) -> Dict[str, float]:
        """
        Extract prosody features from audio

        Args:
            audio: Audio bytes

        Returns:
            Dict of prosody features (pitch, tempo, energy, etc.)
        """
        # TODO: Implement using librosa
        # features = {
        #     "pitch_mean": ...,
        #     "pitch_std": ...,
        #     "tempo": ...,
        #     "energy": ...,
        #     "zero_crossing_rate": ...,
        # }

        return {}

    def _classify_emotion(self, features: Dict[str, float]) -> Emotion:
        """
        Classify emotion from prosody features

        Args:
            features: Prosody features

        Returns:
            Classified emotion
        """
        # TODO: Implement using trained model
        # This could use:
        # - Traditional ML (SVM, Random Forest)
        # - Deep learning (CNN, LSTM)
        # - Pretrained models (wav2vec2, HuBERT)

        return Emotion(name="neutral", confidence=1.0, intensity=0.5)


# Future implementation notes:
"""
Full emotion detection will look like:

import librosa
import numpy as np
from transformers import Wav2Vec2ForSequenceClassification, Wav2Vec2Processor

class EmotionEngine:
    def __init__(self, config: EmotionConfig):
        # Load emotion classification model
        self.model = Wav2Vec2ForSequenceClassification.from_pretrained(
            "ehcalabres/wav2vec2-lg-xlsr-en-speech-emotion-recognition"
        )
        self.processor = Wav2Vec2Processor.from_pretrained(
            "ehcalabres/wav2vec2-lg-xlsr-en-speech-emotion-recognition"
        )

    async def detect_user_emotion(self, audio: bytes) -> Emotion:
        # Convert bytes to numpy array
        audio_array = np.frombuffer(audio, dtype=np.int16).astype(np.float32)

        # Normalize
        audio_array = audio_array / np.max(np.abs(audio_array))

        # Process through model
        inputs = self.processor(
            audio_array,
            sampling_rate=16000,
            return_tensors="pt",
            padding=True
        )

        # Get predictions
        with torch.no_grad():
            logits = self.model(**inputs).logits

        # Get emotion
        predicted_id = torch.argmax(logits, dim=-1).item()
        emotion_name = self.model.config.id2label[predicted_id]
        confidence = torch.softmax(logits, dim=-1).max().item()

        return Emotion(
            name=emotion_name,
            confidence=confidence,
            intensity=confidence  # Could be more sophisticated
        )

Alternative: Use pyannote.audio for emotion detection:

from pyannote.audio import Model, Inference

class EmotionEngine:
    def __init__(self):
        self.model = Model.from_pretrained("pyannote/emotion")
        self.inference = Inference(self.model)

    async def detect_user_emotion(self, audio_file: str) -> Emotion:
        emotion_scores = self.inference(audio_file)
        # emotion_scores is a dict: {"happy": 0.8, "sad": 0.1, ...}

        top_emotion = max(emotion_scores.items(), key=lambda x: x[1])
        return Emotion(
            name=top_emotion[0],
            confidence=top_emotion[1],
            intensity=top_emotion[1]
        )
"""
