extensions/project-soul/soul/memory.py
"""
Memory system using Mem0 for persistent personal context
"""

import asyncio
from typing import List, Dict, Any, Optional
from dataclasses import dataclass
from datetime import datetime
from pathlib import Path
from loguru import logger

from .config import MemoryConfig


@dataclass
class Memory:
    """A single memory entry"""
    id: str
    content: str
    category: str  # "preference", "context", "emotion", "pattern", "project"
    confidence: float
    timestamp: datetime
    metadata: Dict[str, Any]


class SoulMemory:
    """
    Persistent memory system for Soul

    NOTE: This is a placeholder implementation. Full Mem0 integration requires:
    1. Install mem0: pip install mem0ai
    2. Setup vector database (Qdrant or Chroma)
    3. Configure local embeddings
    """

    def __init__(self, config: MemoryConfig, user_id: str):
        self.config = config
        self.user_id = user_id
        self.storage_path = Path(config.storage_path).expanduser()

        logger.info(f"Initializing memory for user: {user_id}")
        logger.warning(
            "⚠️  Mem0 integration is not yet implemented. "
            "This is a placeholder for future implementation."
        )

    async def initialize(self) -> None:
        """Initialize memory system and load existing memories"""
        self.storage_path.mkdir(parents=True, exist_ok=True)

        # TODO: Initialize Mem0 client
        # self.mem0 = Mem0Client(
        #     vector_store=self.config.vector_store,
        #     embedding_model=self.config.embedding_model,
        #     storage_path=str(self.storage_path)
        # )

        logger.info("Memory system initialized")

    async def remember(self, conversation: Dict[str, Any]) -> None:
        """
        Extract and store memories from conversation

        Args:
            conversation: Dict containing user_input, soul_response, emotion, etc.
        """
        # TODO: Implement actual memory storage
        # This should:
        # 1. Extract entities and preferences
        # 2. Categorize information
        # 3. Store in vector database
        # 4. Update knowledge graph

        logger.debug(f"Storing memory: {conversation.get('user_input', '')[:50]}...")

    async def recall(self, query: str) -> List[Memory]:
        """
        Retrieve relevant memories based on query

        Args:
            query: Search query or context

        Returns:
            List of relevant memories
        """
        # TODO: Implement actual memory retrieval
        # This should:
        # 1. Generate query embedding
        # 2. Search vector database
        # 3. Rank by relevance
        # 4. Return top N results

        logger.debug(f"Recalling memories for: {query[:50]}...")
        return []

    async def get_preferences(self, category: Optional[str] = None) -> Dict[str, Any]:
        """
        Get user preferences

        Args:
            category: Optional category filter (e.g., "coding", "ui")

        Returns:
            Dict of preferences
        """
        # TODO: Implement preference retrieval
        # Examples:
        # - "User prefers dark mode"
        # - "User likes Tailwind CSS"
        # - "User prefers detailed explanations"

        return {}

    async def get_context(self, project: Optional[str] = None) -> Dict[str, Any]:
        """
        Get current context and ongoing work

        Args:
            project: Optional project name filter

        Returns:
            Dict of context information
        """
        # TODO: Implement context retrieval
        # Examples:
        # - "Working on React dashboard"
        # - "Current sprint: Authentication refactor"
        # - "Last discussed: Database optimization"

        return {}

    async def get_patterns(self) -> List[Dict[str, Any]]:
        """
        Get recognized user patterns

        Returns:
            List of patterns (work hours, habits, etc.)
        """
        # TODO: Implement pattern recognition
        # Examples:
        # - "Usually codes between 9am-2pm"
        # - "Takes breaks every 90 minutes"
        # - "Prefers voice commands for quick tasks"

        return []

    async def forget(self, memory_id: str) -> None:
        """
        Delete a specific memory (user privacy control)

        Args:
            memory_id: ID of memory to delete
        """
        # TODO: Implement memory deletion
        logger.info(f"Forgetting memory: {memory_id}")

    async def export(self, output_path: Path) -> None:
        """
        Export all memories to file (for backup/portability)

        Args:
            output_path: Path to export file
        """
        # TODO: Implement memory export
        logger.info(f"Exporting memories to: {output_path}")

    async def close(self) -> None:
        """Cleanup resources"""
        logger.info("Closing memory system")
        # TODO: Cleanup database connections


# Future implementation notes:
"""
Full Mem0 integration will look like:

from mem0 import Mem0Client, Memory

class SoulMemory:
    def __init__(self, config: MemoryConfig, user_id: str):
        self.mem0 = Mem0Client(
            vector_store="qdrant",
            embedding_model="all-MiniLM-L6-v2",
            storage_path=config.storage_path
        )
        self.user_id = user_id

    async def remember(self, conversation: dict):
        # Extract entities
        entities = extract_entities(conversation)

        # Store in Mem0
        await self.mem0.add_memories(
            user_id=self.user_id,
            memories=entities,
            metadata={
                "timestamp": datetime.now(),
                "emotion": conversation.get("emotion"),
            }
        )

    async def recall(self, query: str) -> List[Memory]:
        # Search vector database
        results = await self.mem0.search_memories(
            user_id=self.user_id,
            query=query,
            limit=self.config.search_limit
        )
        return results
"""
