extensions/project-soul/soul/moshi_client.py
"""
Moshi voice engine client - handles native speech-to-speech processing
"""

import asyncio
from typing import Optional, Tuple, Any
from dataclasses import dataclass
from loguru import logger

from .config import MoshiConfig


@dataclass
class AudioParams:
    """Parameters for audio output modulation"""
    tone: str = "neutral"
    speed: float = 1.0
    pitch: float = 1.0
    add_pause: bool = False
    add_laughter: bool = False


class MoshiClient:
    """
    Client for Kyutai Moshi native speech-to-speech model

    NOTE: This is a placeholder implementation. Full Moshi integration requires:
    1. Install moshi: pip install moshi-python
    2. Download Moshi model weights: moshi-7b or moshi-3b
    3. Setup audio I/O devices
    """

    def __init__(self, config: MoshiConfig):
        self.config = config
        self.is_listening = False
        self.is_speaking = False

        logger.info(f"Initializing Moshi on {config.device}")
        logger.warning(
            "⚠️  Moshi integration is not yet implemented. "
            "This is a placeholder for future implementation."
        )

    async def listen(self) -> Optional[bytes]:
        """
        Listen for audio input (full duplex)

        Returns:
            Audio bytes if speech detected, None otherwise
        """
        # TODO: Implement actual Moshi listening
        # For now, return None (no input)
        await asyncio.sleep(0.1)
        return None

    async def process(self, audio_input: bytes) -> Tuple[str, dict]:
        """
        Process audio input directly to generate intent and context

        Args:
            audio_input: Raw audio bytes from microphone

        Returns:
            Tuple of (intent_text, context_dict)

        NOTE: Moshi processes audio -> audio natively, but we extract
        semantic intent for task routing
        """
        # TODO: Implement actual Moshi processing
        # This should:
        # 1. Process audio directly (no text transcription)
        # 2. Extract semantic intent
        # 3. Maintain conversation context
        # 4. Detect emotion and urgency

        intent = "placeholder intent"
        context = {
            "confidence": 0.0,
            "detected_emotion": "neutral",
        }

        return intent, context

    async def speak(self, message: str, params: Optional[AudioParams] = None) -> None:
        """
        Generate and play speech output

        Args:
            message: Text to speak (used for semantic content)
            params: Audio modulation parameters
        """
        if params is None:
            params = AudioParams()

        logger.info(f"Soul speaking: {message[:100]}...")

        # TODO: Implement actual Moshi speech generation
        # This should:
        # 1. Generate audio directly from semantic intent
        # 2. Apply emotion modulation (tone, speed, pitch)
        # 3. Stream audio in real-time
        # 4. Support interruption

        self.is_speaking = True
        await asyncio.sleep(len(message) * 0.05)  # Simulate speaking time
        self.is_speaking = False

    def interrupt(self) -> None:
        """Handle user interruption during speaking"""
        if self.is_speaking:
            logger.debug("User interrupted - stopping playback")
            self.is_speaking = False
            # TODO: Implement actual interruption
            # This should:
            # 1. Stop current audio playback immediately
            # 2. Retain partial context
            # 3. Be ready for new input

    async def close(self) -> None:
        """Cleanup resources"""
        logger.info("Closing Moshi client")
        # TODO: Cleanup audio devices, unload model


# Future implementation notes:
"""
Full Moshi integration will look like:

from moshi import MoshiModel, AudioStream

class MoshiClient:
    def __init__(self, config: MoshiConfig):
        self.model = MoshiModel.from_pretrained(
            config.model_path,
            device=config.device
        )
        self.stream = AudioStream(
            sample_rate=config.sample_rate,
            chunk_size=config.chunk_size,
            duplex=config.enable_duplex
        )

    async def listen_and_respond(self):
        async for audio_chunk in self.stream.input:
            # Process audio directly (no text)
            response_audio = await self.model.process(audio_chunk)

            # Stream response immediately
            await self.stream.output.play(response_audio)

            # Extract semantic intent for task routing
            intent = self.model.extract_intent(audio_chunk)
            yield intent
"""
